Conflicts:
	src/Samples/QueryProcessingTest/OffsetkeyTest.cs

+++ src/Starcounter/Query/Optimization/_ExtentNode.cs
@@ -450,7 +450,7 @@ internal class ExtentNode : IOptimizationNode
         return value;
     }
 
-    public Int32 EstimateCost()
+    public Double EstimateCost()
     {
         if (hintedIndexInfo != null || sortIndexInfo != null)
         {
